TTP 300686
The default splash screen now displays a clearer indication that it's loading a game project by using a splash screen that says [Project Nam] Editor.  If we don't have a project name then the old splash screens are used that just says Unreal Editor, or Unreal Engine 4 when launching the game.  For launching the game it will say "Project Name" instead of "Unreal Engine 4".  I've updated the Windows and Mac splash screens to be roughly the same.

This change should help reduce confusion when the editor restarts to load a project, by clearly indicating what is loading.

[CL 2038770 by Nick Darnell in Main branch]