#UE4 Change the Match state machine in GameMode to be explicit and based on the MatchState namespace, as opposed to being based on bools. Refactor related code to be clearer and easier to extend. States:
EnteringMap is the initial state. Actors are not yet ticking and the world hasn't been fully initialized. It will transition to the next state when things are fully loaded.
WaitingToStart (was bWaitingToStartMatch) is the next state, and HandleMatchIsWaitingToStart is called when entering. Actors are ticking, but players have not yet spawned in. It transitions to the next state if ReadyToStartMatch returns true, or if someone calls StartMatch.
InProgress (was bMatchIsInProgress) is the next state, and HandleMatchHasStarted (split off from StartMatch) is called when entering, which calls BeginPlay on all actors. Normal gameplay is in progress. It transitions to the next state if ReadyToEndMatch returns true, or if someone calls EndMatch.
WaitingPostMatch is the next state, and HandleMatchHasEnded (was PerformEndGameHandling) is called when entering. Actors are still ticking but new players are not accepted. It transitions to the next state when a map transfer starts.
LeavingMap (was bLevelChange) is the last state, and HandleLeavingMap (was GameEnding) is called when entering. The match stays in this state while transfering to a new map.
Aborted is the failure state, started from AbortMatch. This is set when there is an unrecoverable error.

Rename functions in GameState and GameSession to match the names above. StartPendingMatch is now HandleMatchIsWaitingToStart, EndPendingMatch/StartMatch is HandleMatchHasStarted, EndGame is HandleMatchHasEnded.
Split the spectator half of functionality of bDelayedStart into a new flag, bStartPlayersAsSpectators. If true, newly joining players will spawn as spectators, if false they'll try to spawn pawns. bDelayedStart now only affects rather the match should start as soon as one player joins, if true the match will not start.
Clean up GameMode::PlayerCanRestart, so it calls PlayerController::CanRestartPlayer, and change all places to call the GameMode. So now if either function returns false the player cannot restart, and games can cleanly override this behavior.
Change GameState::ElapsedTime to only increment when the match has started, the behavior before was inconsistent depending on if you were a client or server
Modify sample games to work with new changes

[CL 2042754 by Ben Zeigler in Main branch]