Can use arrow keys to move through the pin type selector's menu and component's selector menu.
Added a utility widget to help keep focus on a widget and still be able to navigate a listview.

Moved the menu content into a sub-widget that handles the keyboard actions to get the user from the search box to focus on the tree view and handles the enter key.

When using arrow keys in the Pin Type Selector or Component list, item selection will not be acted upon until enter is pressed.

Listviews/Treeviews will no longer skip the first item when using keyboard selection

#ttp 318318 - EDITOR: BP: Variable type selection menu does not support keyboard navigation
#ttp 329065 - BLUEPRINTS: Want to use arrow keys and 'Enter' to select component from the "Add Component" combo box

[CL 2054480 by Michael Schoell in Main branch]