We now use the IMM context provided by Windows which results in the IME state being remembered correctly for UE4
TTP# 331255 - L10N: IME: The user should be able to toggle IME languages and IME needs to remember the state when switching between textboxes

Previously an IMM context was being setup for each Slate context but, since each context manages its own IME state, this resulted in the IME state being reset whenever focus was changed. Even worse was that it didn't even remember the setting for the same Slate widget, because the IMM context was setup again each time it was given focus.

Windows is able to manage the IMM context for each window/thread itself, so I've changed it to do that instead of us handing it manually.
The only caveat is, since we're using Slate, that we need to close the composition ourselves when changing focus since they're not "real" windows (see the call to ImmNotifyIME(IMMContext, NI_COMPOSITIONSTR, CPS_CANCEL, 0))

ReviewedBy Saul.Abreu

[CL 2054533 by Jamie Dale in Main branch]