Altered material editor stats to optionally use relase shaders and to display a "built-in" shader cost.
I've altered the preview material directly to compile with release shaders so that this incurs no overhead.
Refactored the empty material stuff to be neater and to report the base cost of the shader.
Reduced the shaders compiled for this empty stats material to the bare minimum. Typically 3-5, adding a near insignificat cost.

[CL 2054645 by Simon Tovey in Main branch]