Improvements to the way that IMM IME input handles composition sessions
TTP# 332137 - Localization: IME: Google Japanese IME does not handle spacebar correctly

When you first type a character after having pressed space with an IMM IME, you get a WM_IME_COMPOSITION message with both a GCS_RESULTSTR (for the previous composition) and GCS_COMPSTR (for the start of the new composition). I've updated the handling logic to process the result first (and end the current composition session), and then process the new composition after (which starts a new composition session, updating the composition range correctly); this was previously being done the other way around, and resulted in an incorrect composition range.

This change also removes a lot of code that was managing selection state manually, as this isn't required and makes the IMM API calls more in line with the TSF API calls. Selection is now only set in two places:
1) After getting a GCS_RESULTSTR (the caret is moved to the end of the result string).
2) In response to a GCS_CURSORPOS request (this also now honours the CS_NOMOVECARET flag correctly).

Finally, this change fixes issues replacing selected text. Previously if you had text selected and you typed via an IMM IME, your new text would be placed at the start or end of the selection rather than replacing it. Any selected text is now explicitly cleared the first time we get a GCS_COMPSTR, which also further simplifies the handing of the caret position while a composition is running.

ReviewedBy Saul.Abreu

[CL 2056301 by Jamie Dale in Main branch]