Windows XP game support added.
XP support is off by default and currently only supported for win32 builds using Visual Studio 2013
I have updated a lot of platform atomic calls to use 32 bits as 64 is not supported on windows xp.
Added extra checks in d3d11 to ensure that these cached shaders are for that platform.
Added a validity check to the TextInputMethodSystem before we try and process input.
128 bit atomics disabled for header generator. Causes issues with xp compilation. Reviewed by Steve Robb.
Texture Memory size stats changed to from bytes to kilobytes to satisfy dependency for 32 bit atomics.

Note: OpenGL is the default RHI for WindowsXP.

[CL 2057787 by Terence Burns in Main branch]