You can now use an actor picker to choose which actor to attach to
TTP# 279062 - ROCKET: TASK: PUNTABLE: Viewports: "Drag to Attach" feature

Moved FEdModePicker out of the Private section of the PropertyEditor module, and into its own module so that it can be shared between the PropertyEditor and LevelEditor modules. Also renamed it to FEdModeActorPicker.

Added a way to filter FEdModeActorPicker by checking the actor directly (this complements the class type check it already had).

Added an actor picker button next to the scene outliner in the "Attach To" menu on actors. This is also mapped to the Alt+A command.

Built and tested in unity and non-unity.

ReviewedBy Thomas.Sarkanen, Max.Preussner

[CL 2060928 by Jamie Dale in Main branch]