Incorporate support for non-foreign project handling into scheme used by UnrealVersionSelector et al, and automatically register engine installations whenever necessary.
Whenever a blank engine association is read from a .uproject via GetEngineIdentifierForProject(), it checks up the directory hierarchy for an engine capable of loading it via its .uprojectdirs search directories. If found, it will use that engine's local identifier from the registry, or register it if it doesn't exist.

Setting a project's engine association to an engine which is capable of loading it via it's .uprojectdirs will cause it to write a blank engine association string into the .uproject (and triggering the behaviour above on open).

#codereview Michael.Trepka

[CL 2064463 by Ben Marsh in Main branch]