Check for any wrongly saved material coordinates on load
TTP#333911 - Material Editor: Material graph has been reversed

Added another version number to make sure materials aren't continuously checked for this problem in future. Detect whether materials/functions were saved with their coordinates the wrong way around by testing whether most of the inputs to the main outputs are on the right. This will ensure that we don't reverse materials in situations where a strange layout has been setup by the user and we can't be sure whether there is a problem with the coordinates.

[CL 2065439 by Matthew Griffin in Main branch]