Refactoring streaming manager so that it's not so specific to textures.
Renamed FStreamingManagerBase to IStreamingManager and made this into the manager of a Streaming Manager singleton instead of a global variable. Added an interface class to add just the functions specific to texture streaming so that the collection manager doesn't have to deal with textures specifically and prevent this from being part of an Audio Streaming manager in future. This meant a few changes to code using texture functions outside of the streaming managers, with it now having to check that texture streaming is enabled before requesting the texture streaming manager.

[CL 2065528 by Matthew Griffin in Main branch]