Out of date projects now appear faded-out on the project browser
This is to highlight that they will need to be upgraded before they can be used in the current version of the engine. Currently there is no upgrade step, but this will be implemented shortly. This addresses some of TTP #333190 (Project Browser: Improved workflow for opening projects from earlier versions of the editor)

Reviewed by Ben Marsh

[CL 2066916 by Andrew Rodham in Main branch]