Fix for issue introduced by CL 2066955
Details panel would be disabled by default when creating a new blueprint, this was due to checking whether the graph was editable however at that point no graph had been created.

Possible more correct fix would be to editor IsEditable to allow for a null graph, however this would result in changes in several areas of the blueprint editor. In the interest in fixing this issue blocker asap I opted for the more scoped fix to IsFocusedGraphEditable.

The Inspectors now have their scroll bars still enabled when they are disabled too, though this makes them inconsistant with other windows suchs as the pallet in the blueprint editor. These are seperate widgets though and would need addressing independantly.

[CL 2067503 by Barnabas McManners in Main branch]