Fixed TAttribute::Create not working with static function pointers
It was trying to call a constructor which took the function pointer as well as a bool, but it actually seemed to be calling one of the templated constructors instead.

I've simplified the code by removing the constructor which took a function pointer and instead have the Create version which takes a function pointer call FGetter::CreateStatic itself. This simplifies the code and seems to resolve the ambiguity.

ReviewedBy James.Hopkin, Steve.Robb

[CL 2067947 by Jamie Dale in Main branch]