Fixes for how the editor handles localization when the OS is set to a language we don't have a translation for
TTP# 333731 - EDITOR: The editor region doesn't always seem to get set correctly on non-English systems

If you'd previously not set a language in the editor, the "Region & Language" settings would always show you the option for "en_US_POSIX" (and claim that you needed to restart the editor) because the saved culture name ("") didn't match the FInternationalization culture name (eg, "en_GB"), causing it to call GetCulture with an empty string (which seems to cause ICU to return "en_US_POSIX"). It now just uses the current FInternationalization culture name if the saved culture is empty.

Additionally, the editor/game now verifies that it has an available translation for the culture that FInternationalization is using (the one set during startup). If it doesn't (eg, your PC is using "fr_BR") then it will fallback to using generic English ("en") to ensure that the editor/game always has a valid translation to use.

ReviewedBy Saul.Abreu

[CL 2068467 by Jamie Dale in Main branch]