Delegates: Performance optimization for multicast delegates
The main idea is to defer the compaction of the invocation list in order to avoid expensive allocations and deep copies in Broadcast(). Instead of removing bound functions immediately they are merely freed and marked as deleted. The Broadcast() function already iterates over all functions anyway, so the initial check for whether compaction is needed is basically free. Only in the rare cases where compaction is needed, an additional O(n) cost applies. Callees are still able to add new functions, because they will be added to the end of the invocation list and hence ignored by Broadcast() since it is iterating in reverse order. Special guards were added to account for the possibility of re-entrant calls into Broadcast().

#CodeReview: mike.fricker, Richard.TalbotWatkin

[CL 2069229 by Max Preussner in Main branch]