Prevented source control state from being fetched for every content browser widget
Instead of using the (cached) state, we cache the brush to use to display the state and only hit the state cache when state changes.
Also updated Execute() calls to providers to only do relative-path conversions if the path is already relative.
Added SourceControlHelpers::AbsoluteFilenames to aid conversion.

TTP# 335201  - ContentBrowser: Improve idle performance

#codereview Andrew.Brown

[CL 2075346 by Thomas Sarkanen in Main branch]