Fixing crash on shutdown
TTP#336864 - CRITICAL: CRASH: SHOOTERGAME: Shutdown Assertion in FAsyncAduioDecompressWorker::DoWork

Made ICompressedAudioInfo pointer part of the worker and create it in the constructor so that GEngine and Audio Device aren't checked in a different thread. Seems safe to keep using the raw USoundWave pointer as it's been that way for a long time and the decompress worker is checked in USoundWave::IsReadyForFinishDestroy, which is called during garbage collection on shutdown.

[CL 2086960 by Matthew Griffin in Main branch]