XML parser: fixed out-of-bounds memory access and made saving cross-platform by using UTF-8
Local function SplitLines was using BufferSize as if it were the character count (the bug had already been worked around for in-memory XML, but file reading was broken). This happened to work correctly unless it crashed due to the out-of-bounds memory access causing a page fault.

The switch to UTF-8 addresses TTP 336433 (CrashReportClient was silently failing to write the user's id to the crash report XML file).

#CodeReview Bob.Tellez,Thomas.Sarkanen

[CL 2089570 by James Hopkin in Main branch]