Updated the revert window to use Revert Unchanged Only correctly. #TTP 334818 - Editor: Source Control: Revert files dialog checkboxes not behaving properly #branch UE4
#proj Editor.SourceControlWindows

#change Changed the behaviour of the revert window as follows:
Previously:
Clicking revert unchanged only attempted to disable the changed files, all disabling did at this point was to make it seem as if the item was not selected even if it was. This meant that this check box did not fulfil its actual description

Now:
Checking revert unchanged only disables all of the file items as well as the checkbox to enable or disable all of the file items. It then enables the revert button and if pressed it will only send the unchanged files for reversion, this means it fulfills the buttons description.

#change Misc changes: Updated the buttons to use the standardized button layout, and changed "revert files" button to "revert" as this lead to a nicer layout. Added icons to show which of the selected files have been modified.

#reviewedby Thomas.Sarkanen

[CL 2093142 by Barnabas McManners in Main branch]