Fixing Misaligned Material Comments
TTP#337463 - UE4: Samples: Misaligned material nodes need to be tidied up

Noticed that the position of comments was not updated by the scaled size of the comment so it was never going to be moved far enough to the left. Have fixed this for any older materials that still need to be flipped/scaled and added a version number for those inbetween, which will shift the comments left by the desired amount. If the comments have already been lined up manually, they might be forced further out of alignment but there should be less cases of this than those that are actually broken.

[CL 2095637 by Matthew Griffin in Main branch]