Added the concept of a primary asset editor to the asset toolkit manager
The property editor is not considered 'primary' for the purposes of spawning editors. This means that the property editor can be opened, for example, for a static mesh while the static mesh editor is open. Giving focus to both editors when the revlevant editor tries to repon now works correctly. Previously the static mesh editor would give focus to the property editor rather than open up a new mesh editor.

This addresses TTP#334576 - EDITOR: Content Browser: Right clicking on an asset and selecting 'details' will spawn a new details tab even if it is already opened

Reviewed by Tom Sarkanen, Matt Kuhlenschmidt

[CL 2095872 by Andrew Rodham in Main branch]