Improving Opus format so that it has unique identifier
Added a text string at the start of the Opus File format to identify that we're dealing with the right kind of data (discovered this could be a problem when testing the Opus format and was passed OGG data in error).
Added error checking for OGG decompression (Read function could get stuck in an infinite loop if the seek function failed - again only discovered while testing Opus but makes it future proof).

[CL 2100197 by Matthew Griffin in Main branch]