EditorOnlyBP: Function IsInBlueprint needs o check if there is any blueprint. ttp#337444
When CLASS_CompiledFromBlueprint is set and the ClassGeneratedBy is set - the function should return true.
But with EditorOnlyBP enabled the CLASS_CompiledFromBlueprint is set and the ClassGeneratedBy is null. Then the function should return false.
I didnÆt want to add any config setting for EditorOnlyBP outside the editor.ini. Hopefully the function will be removed, once the EditorOnlyBP becomes non-optional.

[CL 2100785 by Maciej Mroz in Main branch]